Extract monthly weather from CF file for input to RothC

Input files need to be named `<in.path>/<in.prefix>.YYYY.nc`

Usage

met2model.RothC(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE
)

Arguments

in.path

path on disk where CF files live

in.prefix

prefix for each file

outfolder

location where model specific output is written.

start_date, end_date

When to start and end output. Specify as exact dates, but output will be padded to whole months.

overwrite

logical: replace output files if they already exist?

Value

data frame summarizing file metadata

Details

Output files are named `<outfolder>/<in.prefix>.YY-mm.YY-mm.dat` with one line per month and columns for temperature, rainfall, and evaporation.

Note that the created file contains only weather data and not any of the soil or management data needed for RothC's single combined input file. See `write.config.RothC()` for assembly into a model-ready RothC_input.dat`.

Author

Chris Black